Hoist shared fixture core_agent_manager
This is used in multiple files so can live in the root conftest.
