+++ tests/conftest.py
@@ -4,13 +4,15 @@ from __future__ import absolute_import, division, print_function, unicode_litera
 import logging
 import os
 import sys
+import subprocess
 
 import pytest
 import wrapt
 from webtest import TestApp
 
 from scout_apm.core import socket as scout_apm_core_socket
-from scout_apm.core.config import SCOUT_PYTHON_VALUES
+from scout_apm.core.core_agent_manager import CoreAgentManager
+from scout_apm.core.config import SCOUT_PYTHON_VALUES, scout_config
 from scout_apm.core.tracked_request import TrackedRequest
 from tests.compat import TemporaryDirectory
 
@@ -109,6 +111,40 @@ def isolate_global_state():
 def core_agent_dir():
     with TemporaryDirectory() as temp_dir:
         yield temp_dir
+
+
+@pytest.fixture
+def core_agent_manager(core_agent_dir):
+    # Shorten path to socket to prevent core-agent from failing with:
+    #   Error opening listener on socket: Custom { kind: InvalidInput,
+    #   error: StringError("path must be shorter than SUN_LEN") }
+    socket_path = "{}/test.sock".format(core_agent_dir)
+    scout_config.set(core_agent_dir=core_agent_dir, socket_path=socket_path)
+    core_agent_manager = CoreAgentManager()
+    try:
+        yield core_agent_manager
+    finally:
+        assert not is_running(core_agent_manager)
+        scout_config.reset_all()
+
+
+def is_running(core_agent_manager):
+    if core_agent_manager.core_agent_bin_path is None:
+        return False
+    agent_binary = [core_agent_manager.core_agent_bin_path, "probe"]
+    socket_path = core_agent_manager.socket_path()
+    probe = subprocess.check_output(agent_binary + socket_path)
+    if b"Agent found" in probe:
+        return True
+    if b"Agent Not Running" in probe:
+        return False
+    raise AssertionError("cannot tell if the core agent is running")
+
+
+def shutdown(core_agent_manager):
+    agent_binary = [core_agent_manager.core_agent_bin_path, "shutdown"]
+    socket_path = core_agent_manager.socket_path()
+    subprocess.check_call(agent_binary + socket_path)
 
 
 # Make all timeouts shorter so that tests exercising them run faster.
