+++ tests/integration/core/test_core_agent_manager.py
@@ -2,49 +2,11 @@
 from __future__ import absolute_import, division, print_function, unicode_literals
 
 import logging
-import subprocess
 import time
 
-import pytest
-
 from scout_apm.core.config import scout_config
-from scout_apm.core.core_agent_manager import CoreAgentManager
 from tests.compat import mock
-
-
-@pytest.fixture
-def core_agent_manager(core_agent_dir):
-    # Shorten path to socket to prevent core-agent from failing with:
-    #   Error opening listener on socket: Custom { kind: InvalidInput,
-    #   error: StringError("path must be shorter than SUN_LEN") }
-    socket_path = "{}/test.sock".format(core_agent_dir)
-    scout_config.set(core_agent_dir=core_agent_dir, socket_path=socket_path)
-    core_agent_manager = CoreAgentManager()
-    try:
-        yield core_agent_manager
-    finally:
-        assert not is_running(core_agent_manager)
-        scout_config.reset_all()
-
-
-def is_running(core_agent_manager):
-    if core_agent_manager.core_agent_bin_path is None:
-        return False
-    agent_binary = [core_agent_manager.core_agent_bin_path, "probe"]
-    socket_path = core_agent_manager.socket_path()
-    probe = subprocess.check_output(agent_binary + socket_path)
-    if b"Agent found" in probe:
-        return True
-    if b"Agent Not Running" in probe:
-        return False
-    raise AssertionError("cannot tell if the core agent is running")
-
-
-def shutdown(core_agent_manager):
-    agent_binary = [core_agent_manager.core_agent_bin_path, "shutdown"]
-    socket_path = core_agent_manager.socket_path()
-    subprocess.check_call(agent_binary + socket_path)
-
+from tests.conftest import is_running, shutdown
 
 # Tests must execute in the order in which they are defined.
